travis: Use the main Dockerfile/docker-compose.yml
Travis CI was using `docker-compose-travis.yml` to load a `Dockerfile`
that was identical to the main one except that PHP dependencies were not
installed in the container but were rather inherited from the
`composer install` call happening outside of it.

If we allow a bit of extra work to happen on Travis CI we can re-use the
main Docker infrastructure by installing PHP dependencies *both* in the
container and on the Travis CI worker itself.